+++ CHANGELOG.md
@@ -12,6 +12,10 @@ All notable changes to this project will be documented in this file.
     - You can disable this on the mod settings.
 - Mod version is now shown in the bottom right corner when pausing.
 - Game Gear unlock messages are no longer shown.
+- Replaced Casino District Keys with Hotel Back Key and Station Back Key.
+- Renamed Hotel/Station Key to Hotel/Station Front Key.
+- Added two new key items: Egglift and Monorail.
+- The inside of the Egg Carrier is now one of the starting locations.
 
 ### Fixed